fixed request type filter + moving things around
