+++ dnsprobe.go
@@ -1,69 +1,110 @@
-package dnsprobe
+package main
 
 import (
-	"net"
+	"bufio"
+	"flag"
+	"fmt"
+	"log"
+	"os"
+	"strings"
 
-	miekdns "github.com/miekg/dns"
-	dns "github.com/projectdiscovery/retryabledns"
+	dnsprobe "github.com/projectdiscovery/dnsprobe/lib"
+	"github.com/remeh/sizedwaitgroup"
 )
 
-// DnsProbe is structure to perform dns lookups
-type DnsProbe struct {
-	dnsClient    *dns.Client
-	questionType uint16
-}
+var (
+	resolvers   = flag.String("s", "", "List of resolvers")
+	hosts       = flag.String("l", "", "List of dns domains")
+	requestType = flag.String("r", "A", "Request Type A, NS, CNAME, SOA, PTR, MX, TXT, AAAA")
+	concurrency = flag.Int("t", 250, "Number of concurrent requests to make")
+	retries     = flag.Int("c", 1, "Max dns retries")
+	verbose     = flag.Bool("v", false, "Output full responses")
+)
 
-// Options contains configuration options
-type Options struct {
-	BaseResolvers []string
-	MaxRetries    int
-	QuestionType  uint16
-}
+func main() {
+	flag.Parse()
 
-// DefaultOptions contains the default configuration options
-var DefaultOptions = Options{
-	BaseResolvers: DefaultResolvers,
-	MaxRetries:    5,
-	QuestionType:  miekdns.TypeA,
-}
+	options := dnsprobe.DefaultOptions
+	options.MaxRetries = *retries
 
-// DefaultResolvers contains the list of resolvers known to be trusted.
-var DefaultResolvers = []string{
-	"1.1.1.1:53",
-	"1.0.0.1:53",
-	"8.8.8.8:53",
-	"8.8.4.4:53",
-}
+	if *resolvers != "" {
+		rs, err := linesInFile(*resolvers)
+		if err != nil {
+			log.Fatalf("Error: %s", err)
+		}
+		options.BaseResolvers = append(options.BaseResolvers, rs...)
+	}
 
-// New creates a dns resolver
-func New(options Options) (*DnsProbe, error) {
-	dnsClient, err := dns.New(options.BaseResolvers, options.MaxRetries)
+	questionType, err := dnsprobe.StringToRequestType(*requestType)
 	if err != nil {
-		return nil, err
+		log.Fatalf("Error: %s", err)
 	}
 
-	return &DnsProbe{dnsClient: dnsClient, questionType: options.QuestionType}, nil
-}
+	options.QuestionType = questionType
 
-// Lookup performs a DNS A question and returns corresponding IPs
-func (d *DnsProbe) Lookup(hostname string) ([]string, error) {
-	if ip := net.ParseIP(hostname); ip != nil {
-		return []string{hostname}, nil
-	}
-
-	results, err := d.dnsClient.Resolve(hostname)
+	var dnsProbe *dnsprobe.DnsProbe
+	dnsProbe, err = dnsprobe.New(options)
 	if err != nil {
-		return nil, err
+		log.Fatalf("Error: %s", err)
 	}
 
-	return results.IPs, nil
+	wg := sizedwaitgroup.New(*concurrency)
+
+	// process file if specified
+	var f *os.File
+	stat, _ := os.Stdin.Stat()
+	// process file if specified
+	if *hosts != "" {
+		var err error
+		f, err = os.OpenFile(*hosts, os.O_RDONLY, os.ModePerm)
+		if err != nil {
+			log.Fatalf("open file error: %v", err)
+			return
+		}
+		defer f.Close()
+	} else if (stat.Mode() & os.ModeCharDevice) == 0 {
+		f = os.Stdin
+	} else {
+		log.Fatalf("hosts file or stdin not provided")
+	}
+
+	sc := bufio.NewScanner(f)
+	for sc.Scan() {
+		wg.Add()
+
+		go func(domain string) {
+			defer wg.Done()
+
+			if ips, err := dnsProbe.LookupRaw(domain); err == nil {
+				for _, ip := range ips {
+					if !(*verbose) {
+						tokens := strings.Split(ip, "\t")
+						ip = tokens[len(tokens)-1]
+					}
+					fmt.Printf("%s %s\n", domain, ip)
+				}
+			}
+		}(sc.Text())
+
+	}
+	if err := sc.Err(); err != nil {
+		log.Printf("scan file error: %v", err)
+		return
+	}
+
+	wg.Wait()
 }
 
-// LookupRaw performs a DNS question of a specified type and returns raw responses
-func (d *DnsProbe) LookupRaw(hostname string) ([]string, error) {
-	if ip := net.ParseIP(hostname); ip != nil {
-		return []string{hostname}, nil
+func linesInFile(fileName string) ([]string, error) {
+	result := []string{}
+	f, err := os.Open(fileName)
+	if err != nil {
+		return result, err
 	}
-
-	return d.dnsClient.ResolveRaw(hostname, d.questionType)
+	scanner := bufio.NewScanner(f)
+	for scanner.Scan() {
+		line := scanner.Text()
+		result = append(result, line)
+	}
+	return result, nil
 }
